Added support for phi-3.5 through Azure
